# design/swagger.yaml
swagger: "2.0"
info:
  version: 0.0.1
  title: Postcard API
  description: Postcard REST API
  contact:
    name: Kari Kangas
    email: karijkangas@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
host: localhost:4000
basePath: /v1
schemes:
  - http
...
paths:
...
  /registrations:
    post:
      summary: Submit registration request
      description: >
        Submit registration request. Service contacts the user via email with instructions how to complete the registration process. On success, returns request expiration time as numbers of milliseconds elapsed since January 1, 1970 00:00:00 UTC. Returns an error if the input data is invalid or the user (email) already exists.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: data
          in: body
          schema:
            $ref: '#/definitions/RegistrationRequest'
      responses:
        202:
          description: Success
          schema:
            $ref: '#/definitions/ExpiringRequestResponse'
        400:
          description: Invalid data 
          schema:
            $ref: '#/definitions/Error'
        409:
          description: User already exists
          schema:
            $ref: '#/definitions/Error'
      tags:
        - requests
...
  /password_resets:
    post:
      summary: Submit password reset request
      description: >
        Submit password reset request. Service contacts the user via email with instructions how to complete the reset process. On success, returns request expiration time as numbers of milliseconds elapsed since January 1, 1970 00:00:00 UTC.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: data
          in: body
          schema:
            $ref: '#/definitions/PasswordResetRequest'
      responses:
        202:
          description: Success
          schema:
            $ref: '#/definitions/ExpiringRequestResponse'
        400:
          description: Invalid data 
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/Error'
      tags:
        - requests
...
  /images:
    post:
      summary: Create upload image
      description: >
        Create upload image. The returned data contains the image id and the URL to upload (PUT) the image data.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: Session token as 'POSTCARD-TOKEN token="1111-2222-3333-4444"'
          required: true
          type: string
      responses:
        201:
          description: Success
          schema:
            $ref: '#/definitions/ImageResponse'
        403:
          description: Invalid session
          schema:
            $ref: '#/definitions/Error'
      tags:
        - images
...
  /users:
    get:
      summary: Find user(s)
      description: >
        Find user(s) matching the input parameters. Returns an error if no input parameters are defined.
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: Session token as 'POSTCARD-TOKEN token="1111-2222-3333-4444"'
          required: true
          type: string
        - name: email
          in: query
          type: string
          description: User email
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/Users'
        400:
          description: Invalid request 
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Invalid session
          schema:
            $ref: '#/definitions/Error'
      tags:
        - users
...
  /users/friends:
    get:
      summary: Get friends
      description: >
        Get friends. Indices are 64-bit integers encoded as strings.
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: Session token as 'POSTCARD-TOKEN token="1111-2222-3333-4444"'
          required: true
          type: string
        - name: exclusiveStartIndex
          in: query
          type: string
          description: Exclusive start index, defaults to "0"
        - name: limit
          in: query
          type: integer
          description: Result limit, defaults to 1000, the max value
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/IndexedUsers'
        400:
          description: Invalid request 
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Invalid session
          schema:
            $ref: '#/definitions/Error'
      tags:
        - users
...
  /postcards:
    post:
      summary: Create a postcard
      description: >
        Create a new postcard. On success, returns the postcard id. The upload image is copied in so it can be removed after the postcard has been created succesfully.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: Session token as 'POSTCARD-TOKEN token="1111-2222-3333-4444"'
          required: true
          type: string
        - name: data
          in: body
          schema:
            $ref: '#/definitions/CreatePostcard'
      responses:
        201:
          description: Success
          schema:
            $ref: '#/definitions/CreatePostcardResponse'
        400:
          description: Invalid data
          schema:
            $ref: '#/definitions/Error'            
        403:
          description: Invalid session, Blocked
          schema:
            $ref: '#/definitions/Error'
        409:
          description: User not found
          schema:
            $ref: '#/definitions/Error'
      tags:
        - postcards
...
  /dev/users:
    post:
      summary: Create user (DEVELOPMENT API)
      description: >
        Create user.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: data
          in: body
          schema:
            $ref: '#/definitions/DevUserData'
      responses:
        201:
          description: Success
          schema:
            $ref: '#/definitions/DevUserDataResponse'
        400:
          description: Invalid data
          schema:
            $ref: '#/definitions/Error'
      tags:
        - dev

    get:
      summary: Get users (DEVELOPMENT API)
      description: >
        Get users.
      produces:
        - application/json
      parameters:
        - name: email
          in: query
          type: string
          description: User email
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/Users'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/Error'
      tags:
        - dev
  /dev/users/{id}:
    delete:
      summary: Delete user (DEVELOPMENT API)
      description: >
        Delete user.
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: User id
      responses:
        204:
          description: Success
      tags:
        - dev
...
  /dev/users/friends:
    post:
      summary: Create friend connection (DEVELOPMENT API)
      description: >
        Create friend connection between two users.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: data
          in: body
          schema:
            $ref: '#/definitions/DevFriends'
      responses:
        201:
          description: Success
          schema:
            $ref: '#/definitions/DevConnection'
        400:
          description: Invalid data
          schema:
            $ref: '#/definitions/Error'
      tags:
        - dev
...
  /dev/ignored:
    post:
      summary: Add email to Do Not Disturb list (DEVELOPMENT API)
      description: >
        Add email to Do Not Disturb list.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: data
          in: body
          schema:
            $ref: '#/definitions/DevEmail'
      responses:
        204:
          description: Success
        400:
          description: Invalid data
          schema:
            $ref: '#/definitions/Error'
      tags:
        - dev

    get:
      summary: Get ignored list (DEVELOPMENT API)
      description: >
        Get ignored list.
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/DevIgnored'
      tags:
        - dev
...
  /dev/postcards:
    get:
      summary: Get postcards (DEVELOPMENT API)
      description: >
        Get postcards.
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/GetPostcardsResponse'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/Error'
      tags:
        - dev
  /dev/postcards/{id}:
    delete:
      summary: Delete postcard (DEVELOPMENT API)
      description: >
        Delete postcard.
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: Postcard id
      responses:
        204:
          description: Success
      tags:
        - dev
...
definitions:
...
  RegistrationRequest:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
        format: password
      language:
        type: string
    required:
      - firstName
      - lastName
      - email
      - password
      - language
  ExpiringRequestResponse:
    type: object
    properties:
      expires:
        type: integer
        format: int64
        example: 1553028728606
...
  PasswordResetRequest:
    type: object
    properties:
      email:
        type: string
    required:
      - email
...
  User:
    type: object
    properties:
      id:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      avatar:
        type: string
  Users:
    type: array
    items:
      $ref: '#/definitions/User'
  IndexedUser:
    type: object
    properties:
      index:
        type: string
      id:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      avatar:
        type: string
  IndexedUsers:
    type: array
    items:
      $ref: '#/definitions/IndexedUser'
...
  CreatePostcard:
    type: object
    properties:
      receiver:
        type: string
      image:
        type: string
      message:
        type: string
      location:
        type: string
    required:
      - receiver
      - image
      - location
  CreatePostcardResponse:
    type: object
    properties:
      id:
        type: string
...
  Postcard:
    type: object
    properties:
      id:
        type: string
      sender:
        type: string
      receiver:
        type: string
      image:
        type: string
      message:
        type: string
      location:
        type: string
      created:
        type: string
        format: date-time
      read:
        type: string
        format: date-time
  GetPostcardsResponse:
    type: array
    items:
      $ref: '#/definitions/Postcard'
...
  ImageResponse:
    type: object
    properties:
      id:
        type: string
      url:
        type: string
    required:
      - id
      - url
  Error:
    type: object
    properties:
      error:
        type: string
...
  DevUserData:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
      language:
        type: string
  DevFriends:
    type: object
    properties:
      user:
        type: string
      friend:
        type: string
  DevConnection:
    type: object
    properties:
      index:
        type: integer
        format: int64
      user:
        type: string
      friend:
        type: string
  DevUserDataResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      passhash:
        type: string
      language:
        type: string
      avatar:
        type: string
...
  DevEmail:
    type: object
    properties:
      email:
        type: string
  DevIgnored:
    type: object
    properties:
      hashes:
        type: array
        items:
          type: string